services/django/requirements.txt
Django
djangorestframework
django-extensions
aiohttp[speedups]
aioresponses
spotipy2

# Test
factory_boy
faker_music
pytest-django
pytest-asyncio
pytest-bdd
